requirements.txt
pandas==2.2.2
numpy==1.26.4
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
rank-bm25==0.2.2
scikit-learn==1.4.2
tqdm==4.66.4
streamlit==1.37.1
plotly==5.22.0
matplotlib==3.8.4
seaborn==0.13.2
sqlalchemy==2.0.30
tabulate==0.9.0
python-dotenv==1.0.1
openai==1.40.6
httpx==0.27.0
